Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -5,6 +5,6 @@ int main(){
     float b;
     scanf("%f",&a);
     scanf("%f",&b);
-    double product = a*b;
-    printf("Product: %lf",product);
+    float product = a*b;
+    printf("Product: %f",product);
 }
